skill-bridge/requirements.txt
fastapi==0.135.1
httpx==0.28.1
protobuf==7.34.0
pydantic==2.12.5
pytest==9.0.2
python-dotenv==1.2.2
